@@ -12,8 +12,8 @@
     "@dietime/react-native-date-picker": "^1.2.1",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.14.16",
-    "@mui/material": "^5.14.17",
+    "@mui/icons-material": "^5.14.18",
+    "@mui/material": "^5.14.18",
     "@mui/styled-engine-sc": "^5.14.12",
     "@react-spring/web": "^9.7.3",
     "css-mediaquery": "^0.1.2",
@@ -23,7 +23,7 @@
     "moment": "^2.29.4",
     "moment-locales-webpack-plugin": "^1.2.0",
     "moment-timezone-data-webpack-plugin": "^1.5.1",
-    "next": "14.0.2",
+    "next": "14.0.3",
     "next-auth": "^4.22.3",
     "query-string": "^8.1.0",
     "react": "18.2.0",
@@ -35,12 +35,12 @@
     "sharp": "^0.32.6",
     "styled-components": "^5.3.11",
     "swiper": "^9.4.1",
-    "typescript": "^5.2.2",
+    "typescript": "^5.3.2",
     "ua-parser-js": "^1.0.37",
     "zustand": "^4.4.6"
   },
   "devDependencies": {
-    "eslint": "8.53.0",
-    "eslint-config-next": "14.0.2"
+    "eslint": "8.54.0",
+    "eslint-config-next": "14.0.3"
   }
 }
